Override callback_url method provided by OmniAuth::Strategies
omniauth-oauth2 1.4.0 removed overriding this method itself - this
caused an issue where the callback_url was not exactly matching what the
provider (dotloop) expects and returning an error during the oauth
authentication process.

The omniauth gem appends any `querystring` value to the callback_url
which is sort of OAuth 2.0 spec compliant but the spec is vague so
having the querystring in the callback url isn't hard defined - hence
the differences between providers.

In any case, dotloop wants no querystring in the callback url so we
override the method here to remove the querystring